Add files via upload
Fire incidents are becoming one of the most urgent global issues. According to recent studies, the cost of fires in developed countries is estimated to reach 1% of the GDP amounting to around \$76bn worldwide (Universitas21, no date). Looking at specific countries, the estimated economic lost due to fire incidents in UK between 2017 and 2018 was between £5.73bn and £11.46bn (NFCC, 2019), while in 2016, the economic loss in Canada on cost of fires were up to $7.6 billion (Douglas, 2016). This project is to focus on the estimated economic loss on fire of Canada’s largest city, Toronto, due to the recent news that the city is facing an understaffed and tight budget for fire issues, especially during periods of COVID restrictions when residents were encouraged to spend more time at home (Boisvert, 2021). There were 10% of fire trucks unavailable due to insuﬀicient firefighters; according to the Toronto Professional Fire Fighters Association (TPFFA). Our target is to know the estimated economic loss on fire as a reference for the government’s budgeting.
By considering how much money the government should budget for the fire department, we believe it is significant to know the loss of economic cost fire incidents can cause. Therefore, we will answer the following question: What is the estimated economic loss of a fire incident?

Data provided by the City of Toronto