Add files via upload
The Red Wine Quality dataset chosen for this project is a popular dataset for machine learning research and education. It is available on both the UCI Machine Learning Repository and Kaggle, and contains 1599 instances and 12 attributes. The attributes include various chemical properties of red wine, such as fixed acidity, volatile acidity, and alcohol content, as well as the quality rating of the wine. The goal of this project is to predict the quality rating of red wine based on its chemical properties, using MLP multi-class classification. The quality rating is a categorical variable with six levels, ranging from 1 as the worst to 6 as the best quality. By applying MLP multi-class classification to this dataset, the target is to develop a model that can accurately predict the quality of red wine based on its chemical properties, which can be useful in the wine industry and for wine connoisseurs. The reason to choose multi-class classification over binary classification is that the decision between good and bad quality should be decided by the wine connoisseurs instead of simply splitting the quality to two classes.